Add .gitignore
Make sure to always include a .gitignore in the
root of your node projects repo/directory.

This is a plain text file that takes in plain
text entries of filenames and directories to
ignore with git.

We want to exclude the ‘node_modules’ folder
since we can install the most up-to-date
dependencies with only a package.json and
`npm install`